+++ biconnected.hpp
@@ -12,15 +12,15 @@ class FaceTriangulation;
 class biconnected
 {
 public:
-    vector<vector<long long>> faces;
-    unordered_set<pair<long long, long long>, PairHash> present;
-    vector<vector<pair<long long, long long>>> allTriangulations;
+    vector<vector<int>> faces;
+    unordered_set<pair<int, int>, PairHash> present;
+    vector<vector<pair<int, int>>> allTriangulations;
     vector<FaceTriangulation*> faceTriangulations;
     u128 totalTriangulations = 0;
-    biconnected(vector<vector<long long>> &faces)
+    biconnected(vector<vector<int>> &faces)
     {
         this->faces = faces;
-        present = unordered_set<pair<long long, long long>, PairHash>();
+        present = unordered_set<pair<int, int>, PairHash>();
         initiatePresent();
         faceTriangulations = vector<FaceTriangulation*>(faces.size());
     }
@@ -28,7 +28,7 @@ public:
     {
         for (auto face : faces)
         {
-            for (long long i = 0; i < face.size() - 1; i++)
+            for (int i = 0; i < face.size() - 1; i++)
             {
                 present.insert(make_pair(min(face[i], face[i + 1]), max(face[i], face[i + 1])));
             }
@@ -36,7 +36,7 @@ public:
         }
     }
     void getAllTriangulations();
-    void output(long long serial);
+    void output(int serial);
     void addTriangulation();
     void sortTriangulations()
     {
@@ -68,7 +68,7 @@ inline void biconnected::getAllTriangulations()
     // printAllTriangulations();
 }
 
-inline void biconnected::output(long long serial)
+inline void biconnected::output(int serial)
 {
     if (serial == faces.size() - 1)
     {
@@ -87,10 +87,10 @@ inline void biconnected::output(long long serial)
 /// @brief Adds the current triangulation (combination of triangulations from all faces) to the list of all triangulations in the biconnected component.
 inline void biconnected::addTriangulation()
 {
-    vector<pair<long long, long long>> currentTriangulations;
+    vector<pair<int, int>> currentTriangulations;
     for (auto a : faceTriangulations)
     {
-        vector<pair<long long, long long>> currentTriangulation;
+        vector<pair<int, int>> currentTriangulation;
         for (auto &chord : a->chords)
         {
             currentTriangulation.push_back(a->getPair(chord));
